Move color to columns package
