+++ server/src/columns/database_dto.go
@@ -3,7 +3,6 @@ package columns
 import (
 	"github.com/google/uuid"
 	"github.com/uptrace/bun"
-	"scrumlr.io/server/database/types"
 )
 
 // Column the model for a column of a board
@@ -13,7 +12,7 @@ type DatabaseColumn struct {
 	Board         uuid.UUID
 	Name          string
 	Description   string
-	Color         types.Color
+	Color         Color
 	Visible       bool
 	Index         int
 }
@@ -24,7 +23,7 @@ type DatabaseColumnInsert struct {
 	Board         uuid.UUID
 	Name          string
 	Description   string
-	Color         types.Color
+	Color         Color
 	Visible       *bool
 	Index         *int
 }
@@ -36,7 +35,7 @@ type DatabaseColumnUpdate struct {
 	Board         uuid.UUID
 	Name          string
 	Description   string
-	Color         types.Color
+	Color         Color
 	Visible       bool
 	Index         int
 }
